Move logic to common
